Preliminary HDDMRegressor Analysis
mw_ddm has the code for the basic model that I have gotten working with HDDMRegressor for the trial-by-trial regression coefficients for slow-waves on decision making parameters. 
At this stage this means that the data are stimulus coded on v as opposed to z but I am working on making it possible for z too, (in practice this doesn't matter a whole lot but all my previous work I've stimulus coded on z so I prefer that. 

the wfpt.pyx is a source file that you will need to replace in your current version of HDDM. So, you will need to uninstall your HDDM, replace the file, and then re-install. The change makes it possible to use dummy codes for No-Go responses in HDDMRegressor (its the same change as is made in other sections of the code that allow it to happen in HDDMStimCoding). Dummy coded responses are coded as -1. 

Also, please note that I have removed RTs <.2s and greater than 1s. from the data and you may need to rename a couple of your columns so they work with HDDM.